# shape: move calculate-svg-dimensions require where it's used