--- a/lib/svg-sprite/shape.js
+++ b/lib/svg-sprite/shape.js
@@ -18,7 +18,6 @@ const xpath = require('xpath');
 const cssom = require('cssom');
 const { CssSelectorParser } = require('css-selector-parser');
 const async = require('async');
-const calculateSvgDimensions = require('./calculate-svg-dimensions.js');
 const fixXMLString = require('./fix-xml-string.js');
 
 /**
@@ -535,6 +534,9 @@ SVGShape.prototype._determineDimensions = function(cb) {
 
     // If the viewBox attribute didn't suffice: Render the SVG image
     if (!this.width || !this.height) {
+        // We require this here so that we don't increase module load time
+        const calculateSvgDimensions = require('./calculate-svg-dimensions.js');
+
         try {
             const dimension = calculateSvgDimensions(this.getSVG(false));
             this.height = this._round(dimension.height);